Filtering books now works. All basic functions are implemented.

+++ View/books.php
@@ -1,42 +1,59 @@
 <?php
 include_once 'navbar.php';
 $output ="";
+$books ="";
 if ($_SERVER["REQUEST_METHOD"] == "POST") {
-    $books = $_POST['bookcbs'];
-    if (count($books) > 0) {
-        include_once '../DAL/DB.php';
+    if (isset($_POST['delete'])) {
+        $books = $_POST['bookcbs'];
+        if (count($books) > 0) {
+            include_once '../DAL/DB.php';
 
-        $db = new DB();
+            $db = new DB();
 
-        $output = $db->deleteBooks($books);
-    } else {
-        $output = "No book(s) selected.";
+            $output = $db->deleteBooks($books);
+        } else {
+            $output = "No book(s) selected.";
+        }
     }
-    
+    if (isset($_POST['searchButton'])) {
+        include_once '../DAL/DB.php';
+        $db = new DB();
+        if (!empty($_POST['searchText'])) {
+            $filter = $_POST['searchText'];
+            $books = $db->displayBooks($filter);
+        } else {
+            $books = $db->displayAllBooks();
+        }
+    }
 }
 ?>
 
 
-
-<form action="<?php echo $_SERVER['PHP_SELF'];?>" method="post">
+<div class="main-wrapper">
 <h1>List of all library books</h1>
 <div class="search-wrapper">
     <form action="<?php echo $_SERVER['PHP_SELF'];?>" method="post">
-        <input type="text" name="searchText" placeholder="Filter by title, author, etc..." required>
-        <input type="button" name="searchButton" value="Filter">
+        <input type="text" name="searchText" placeholder="Filter by title, author, etc..."\>
+        <input type="submit" name="searchButton" value="Filter">
     </form>
 </div>
+<form action="<?php echo $_SERVER['PHP_SELF'];?>" method="post">
 <?php
 include_once '../DAL/DB.php';
+if ($books == "") {
+    $db = new DB();
 
-$db = new DB();
-
-$books = $db->displayAllBooks();
+    $books = $db->displayAllBooks();
+    echo $books;
+} else {
+    echo $books;
+}
 ?>
-<input type="submit" name="delete" value="Delete selected book(s)">
+<input id="deleteButton" type="submit" name="delete" value="Delete selected book(s)">
 </form>
 <?php
 echo $output;
 ?>
+</div>
